feat: 컨트롤러 타입 적용, 타입가드 추가

+++ mission-4/src/controllers/productController.ts
@@ -1,3 +1,4 @@
+import type { RequestHandler } from 'express';
 import {
   getProductService,
   getProductListService,
@@ -7,46 +8,73 @@ import {
   postProductLikeService,
   deleteProductLikeService,
 } from '../services/productService.js';
+import { hasId, hasIdAndUserId, hasParsedQuery, hasTokenPayload } from '@/types/guard.js';
+import { BadRequestError } from '@/lib/errors.js';
 
 class ProductController {
-  async getProduct(req, res) {
+  getProduct: RequestHandler = async (req, res) => {
+    if (!hasId(req)) {
+      throw new BadRequestError();
+    }
     const { id: productId } = req.parsedId;
-    const { id: userId } = req.tokenPayload || {};
+    const { userId } = req.tokenPayload || {};
     const result = await getProductService({ userId, productId });
     return res.status(200).json(result);
-  }
-  async getProductList(req, res) {
-    const { id: userId } = req.tokenPayload || {};
+  };
+  getProductList: RequestHandler = async (req, res) => {
+    if (!hasParsedQuery(req)) {
+      throw new BadRequestError();
+    }
+    const { userId } = req.tokenPayload || {};
     const result = await getProductListService({ userId, ...req.parsedQuery });
     return res.status(200).json(result);
-  }
-  async postProduct(req, res) {
-    const { id: userId } = req.tokenPayload;
+  };
+  postProduct: RequestHandler = async (req, res) => {
+    if (!hasTokenPayload(req)) {
+      throw new BadRequestError();
+    }
+    // const { id: userId } = req.tokenPayload;
+    const { userId } = req.tokenPayload;
     const result = await postProductService({ userId, ...req.body });
     return res.status(201).json(result);
-  }
-  async patchProduct(req, res) {
-    const id = req.parsedId;
-    const data = req.body;
-    const result = await patchProductService({ id, data });
-    return res.status(200).json(result);
-  }
-  async deleteProduct(req, res) {
-    const result = await deleteProductService({ id: req.parsedId });
-    return res.status(200).json(result);
-  }
-  async postProductLike(req, res) {
+  };
+  patchProduct: RequestHandler = async (req, res) => {
+    if (!hasIdAndUserId(req)) {
+      throw new BadRequestError();
+    }
     const { id: productId } = req.parsedId;
-    const { id: userId } = req.tokenPayload;
+    const { userId } = req.tokenPayload;
+    const data = req.body;
+    const result = await patchProductService({ userId, productId, data });
+    return res.status(200).json(result);
+  };
+  deleteProduct: RequestHandler = async (req, res) => {
+    if (!hasIdAndUserId(req)) {
+      throw new BadRequestError();
+    }
+    const { id: productId } = req.parsedId;
+    const { userId } = req.tokenPayload;
+    const result = await deleteProductService({ userId, productId });
+    return res.status(200).json(result);
+  };
+  postProductLike: RequestHandler = async (req, res) => {
+    if (!hasIdAndUserId(req)) {
+      throw new BadRequestError();
+    }
+    const { id: productId } = req.parsedId;
+    const { userId } = req.tokenPayload;
     const result = await postProductLikeService({ userId, productId });
     return res.status(201).json(result);
-  }
-  async deleteProductLike(req, res) {
+  };
+  deleteProductLike: RequestHandler = async (req, res) => {
+    if (!hasIdAndUserId(req)) {
+      throw new BadRequestError();
+    }
     const { id: productId } = req.parsedId;
-    const { id: userId } = req.tokenPayload;
+    const { userId } = req.tokenPayload;
     const result = await deleteProductLikeService({ userId, productId });
     return res.status(200).json(result);
-  }
+  };
 }
 
 export default new ProductController();
